requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
astroid==3.3.5
bcrypt==4.2.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
click==8.1.7
cryptography==43.0.1
dill==0.3.9
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
factory_boy==3.3.1
Faker==30.1.0
fastapi==0.115.0
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
identify==2.6.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy==1.11.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pylint==3.3.1
pytest==8.3.3
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.6
watchfiles==0.24.0
websockets==13.1
